requirements/fastapi.txt
uvicorn==0.18.2
mlflow==2.9.1
pandas==1.5.3
numpy==1.24.4
lz4==4.3.2
psutil==5.9.5
scipy==1.14.1
Jinja2==3.1.2
python-multipart==0.0.6
scikit-learn==1.6.1
langchain==0.3.16
langchain-community==0.3.16
langchain-core==0.3.32
langchain-text-splitters==0.3.5
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
yfinance==0.2.52
google-generativeai==0.8.4
beautifulsoup4==4.12.3
fastapi==0.110.0
pydantic==2.7.4
ta==0.11.0
grpcio==1.70.0
grpcio-status==1.62.3
requests==2.32.3
httpx==0.27.0
grpcio-tools==1.70.0
pytest==8.2.0
pytest-asyncio==0.23.7
mypy==1.10.0
black==24.4.2
sqlalchemy==2.0.37
psycopg2-binary==2.9.9
loguru==0.7.2
rich==13.7.1
pyarrow==14.0.2
h11==0.14.0
google-api-python-client==2.108.0
google-auth==2.27.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
click==7.0
typing_extensions==4.12.2
packaging==24.2
urllib3==2.3.0
idna==3.10
certifi==2024.12.14
msgpack==1.1.0
tenacity==9.0.0
langsmith==0.3.2
requests_toolbelt==1.0.0
sniffio==1.3.1
httpcore==1.0.7
zstandard==0.23.0
jsonpatch==1.33
jsonpointer==3.0.0
annotated-types==0.4.0
python-dateutil==2.8.2
pytz==2024.2
six==1.17.0
httplib2==0.22.0
pyparsing==3.2.1
uritemplate==4.1.1
googleapis-common-protos==1.66.0
google-api-core==2.19.2
googledrivedownloader==0.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
rsa==4.9
cachetools==5.5.1
google-ai-generativelanguage==0.6.15
protobuf==4.25.6
proto-plus==1.26.0
pydantic-core==2.27.2
tqdm==4.67.1
PyYAML==6.0.2
databricks-cli==0.18.0
sqlparse==0.5.3
entrypoints==0.4
importlib_metadata==8.6.1
starlette
anyio==3.7.1
pyjwt==1.7.0
tabulate==0.7.7
aiohttp==3.8.3
chardet==5.2.0
charset-normalizer==3.4.1
cloudpickle==3.1.1
joblib==1.4.2
threadpoolctl==3.5.0
feedparser==6.0.11
sgmllib3k==1.0.0
gitpython==3.1.44
gitdb
chroma-hnswlib==0.7.6
chromadb==0.4.22
overrides==7.4.0
onnxruntime==1.20.1
tokenizers==0.21.0
opentelemetry-api ==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
deprecated==1.2.18
wrapt==1.17.2
posthog==3.16.0
distro
backoff==2.2.1
monotonic==1.6
orjson==3.10.15
